chore: moved type files to type folder
