+++ packages/data-layer/src/index.ts
@@ -1,6 +1,6 @@
 export { DataLayer } from "./data-layer";
 export { useDataLayer, DataLayerContext, DataLayerProvider } from "./react";
 export * from "./openapi-search-client/models/index";
-export * from "./data.types";
-export * from "./roundApplication.types";
+export * from "./types/data.types";
+export * from "./types/roundApplication.types";
 export * from "./utils";
